Fixed the semantic of ADD, OR and XOR
Furthermore, the corresponding tests for each operator have also been added

+++ src/main/java/it/unipr/analysis/EVMAbsDomain.java
@@ -490,8 +490,8 @@ public class EVMAbsDomain implements ValueDomain<EVMAbsDomain>, BaseLattice<EVMA
 					MathNumber low, high;
 
 					try {
-						low = new MathNumber(opnd1.interval.getLow().toByte() & opnd2.interval.getLow().toByte());
-						high = new MathNumber(opnd1.interval.getHigh().toByte() & opnd2.interval.getHigh().toByte());
+						low = new MathNumber(opnd1.interval.getLow().toLong() & opnd2.interval.getLow().toLong());
+						high = new MathNumber(opnd1.interval.getHigh().toLong() & opnd2.interval.getHigh().toLong());
 					} catch (MathNumberConversionException e) {
 						result.push(Interval.TOP);
 						return new EVMAbsDomain(result, memory, mu_i);
@@ -510,8 +510,8 @@ public class EVMAbsDomain implements ValueDomain<EVMAbsDomain>, BaseLattice<EVMA
 					MathNumber low, high;
 
 					try {
-						low = new MathNumber(opnd1.interval.getLow().toByte() | opnd2.interval.getLow().toByte());
-						high = new MathNumber(opnd1.interval.getHigh().toByte() | opnd2.interval.getHigh().toByte());
+						low = new MathNumber(opnd1.interval.getLow().toLong() | opnd2.interval.getLow().toLong());
+						high = new MathNumber(opnd1.interval.getHigh().toLong() | opnd2.interval.getHigh().toLong());
 					} catch (MathNumberConversionException e) {
 						result.push(Interval.TOP);
 						return new EVMAbsDomain(result, memory, mu_i);
@@ -530,8 +530,8 @@ public class EVMAbsDomain implements ValueDomain<EVMAbsDomain>, BaseLattice<EVMA
 					MathNumber low, high;
 
 					try {
-						low = new MathNumber(opnd1.interval.getLow().toByte() ^ opnd2.interval.getLow().toByte());
-						high = new MathNumber(opnd1.interval.getHigh().toByte() ^ opnd2.interval.getHigh().toByte());
+						low = new MathNumber(opnd1.interval.getLow().toLong() ^ opnd2.interval.getLow().toLong());
+						high = new MathNumber(opnd1.interval.getHigh().toLong() ^ opnd2.interval.getHigh().toLong());
 					} catch (MathNumberConversionException e) {
 						result.push(Interval.TOP);
 						return new EVMAbsDomain(result, memory, mu_i);
